requirements.txt
flask
flask-cors
python-dotenv
openai==0.28
